% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get4ceFile}
\alias{get4ceFile}
\title{Import a 4CE CSV file}
\usage{
get4ceFile(
  filename,
  dir.input = getInputDataDirectoryName(),
  fread = getOption("4ce.fread", FALSE),
  tolower = TRUE,
  ...
)
}
\arguments{
\item{filename}{Name of the file, such as LocalPatientSummary.csv}

\item{dir.input}{Input directory, which is via \code{\link[=getInputDataDirectoryName]{getInputDataDirectoryName()}}
if using docker.}

\item{fread}{if TRUE, use data.table::fread() for faster import. Otherwise
use read.csv(). This will also check the 4ce.fread option to make it work
without passing a function argument.}

\item{tolower}{Lowercase column names? (default: yes)}

\item{...}{(Optional) Additional arguments passed through to fread() or
read.csv()}
}
\description{
Files are loaded from the input data directory, optionally using \code{\link[data.table:fread]{data.table::fread()}}.
}
\seealso{
\code{\link[data.table:fread]{data.table::fread()}}, \code{\link[utils:read.table]{utils::read.csv()}}
}
