% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileFunctions.R
\name{get4ceRootDirectoryName}
\alias{get4ceRootDirectoryName}
\title{Returns the name of the root of the file system location to be used as workspace.}
\usage{
get4ceRootDirectoryName()
}
\description{
Returns the name of the root of the file system location to be used as workspace.
}
\keyword{4CE}
