% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getClinicalCourse}
\alias{getClinicalCourse}
\title{Import ClinicalCourse-yoursite.csv}
\usage{
getClinicalCourse(siteid, ...)
}
\arguments{
\item{siteid}{Name of your site.}

\item{...}{(Optional) Additional arguments passed to \code{\link[=get4ceFile]{get4ceFile()}}.}
}
\description{
Import ClinicalCourse-yoursite.csv
}
