% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerStorageFunctions.R
\name{getContainerScratchDirectory}
\alias{getContainerScratchDirectory}
\title{Returns the name of the directory in the container that packages can use for 
temporary storage (e.g., to persist analysis results between calculation and 
upload).}
\usage{
getContainerScratchDirectory()
}
\arguments{
\item{fullyQualified}{Should the return value be a fully-qualified path?  If not, just the name of the directory is returned.}
}
\description{
Returns the name of the directory in the container that packages can use for 
temporary storage (e.g., to persist analysis results between calculation and 
upload).
}
\keyword{4CE}
