% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getDailyCounts}
\alias{getDailyCounts}
\title{Import DailyCounts-yoursite.csv}
\usage{
getDailyCounts(siteid, ...)
}
\arguments{
\item{siteid}{Name of your site.}

\item{...}{(Optional) Additional arguments passed to \code{\link[=get4ceFile]{get4ceFile()}}.}
}
\description{
Import DailyCounts-yoursite.csv
}
