% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getDemographics}
\alias{getDemographics}
\title{Import Demographics-yoursite.csv}
\usage{
getDemographics(siteid, ...)
}
\arguments{
\item{siteid}{Name of your site.}

\item{...}{(Optional) Additional arguments passed to \code{\link[=get4ceFile]{get4ceFile()}}.}
}
\description{
Import Demographics-yoursite.csv
}
