% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getDiagnoses}
\alias{getDiagnoses}
\title{Import Diagnoses-yoursite.csv}
\usage{
getDiagnoses(siteid, ...)
}
\arguments{
\item{siteid}{Name of your site.}

\item{...}{(Optional) Additional arguments passed to \code{\link[=get4ceFile]{get4ceFile()}}.}
}
\description{
Import Diagnoses-yoursite.csv
}
