% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileFunctions.R
\name{getInputDataDirectoryName}
\alias{getInputDataDirectoryName}
\title{Returns the name of the directory where site input data is expected to reside.}
\usage{
getInputDataDirectoryName(fullyQualified = TRUE)
}
\arguments{
\item{fullyQualified}{Should the return value be a fully-qualified path?  If not, just the name of the directory is returned.}
}
\description{
Returns the name of the directory where site input data is expected to reside.
}
\keyword{4CE}
