% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_nodocker.R
\name{getLocalPatientClinicalCourse_nodocker}
\alias{getLocalPatientClinicalCourse_nodocker}
\title{Import LocaLPatientClinicalCourse.csv}
\usage{
getLocalPatientClinicalCourse_nodocker(siteid, dir.input, ...)
}
\arguments{
\item{siteid}{Name of your site.}

\item{dir.input}{Input directory}

\item{...}{(Optional) Additional arguments passed to \code{\link[=get4ceFile]{get4ceFile()}}.}
}
\description{
Uses \code{\link[=get4ceFile]{get4ceFile()}} - see that file for additional customization.
}
\seealso{
\code{\link[=get4ceFile]{get4ceFile()}}
}
