% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getLocalPatientObservations}
\alias{getLocalPatientObservations}
\title{Import LocalPatientObservations.csv}
\usage{
getLocalPatientObservations(siteid = NULL, ...)
}
\arguments{
\item{siteid}{(unused)}

\item{...}{(Optional) Additional arguments passed to \code{\link[=get4ceFile]{get4ceFile()}}.}
}
\description{
Uses \code{\link[=get4ceFile]{get4ceFile()}} - see that file for additional customization.
}
\seealso{
\code{\link[=get4ceFile]{get4ceFile()}}
}
