% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getMedications}
\alias{getMedications}
\title{Import Medications-yoursite.csv}
\usage{
getMedications(siteid, ...)
}
\arguments{
\item{siteid}{Name of your site.}

\item{...}{(Optional) Additional arguments passed to \code{\link[=get4ceFile]{get4ceFile()}}.}
}
\description{
Uses \code{\link[=get4ceFile]{get4ceFile()}} - see that file for additional customization.
}
\seealso{
\code{\link[=get4ceFile]{get4ceFile()}}
}
