% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFileFunctions.R
\name{getSiteId}
\alias{getSiteId}
\title{Returns the name of the site whose files are in the directory returned by getInputDataDirectoryName().
Throws an error if files are inconsistently named, or if files for multiple site ids are present,
or if required files are missing.}
\usage{
getSiteId()
}
\description{
Returns the name of the site whose files are in the directory returned by getInputDataDirectoryName().
Throws an error if files are inconsistently named, or if files for multiple site ids are present,
or if required files are missing.
}
\keyword{4CE}
